+++ packages/react-components/src/Scenarios/ScenarioItem/ScenarioItem.tsx
@@ -18,6 +18,7 @@ const ScenarioItem = (props: ScenarioItemOLDProps) => {
     name,
     description,
     showMenu,
+    onClick,
     onContextMenuClick,
     menu,
     scenario,
@@ -76,17 +77,27 @@ const ScenarioItem = (props: ScenarioItemOLDProps) => {
   );
 
   return (
-    <div
-      className={classes.scenario}
-      onMouseOver={() => setHover(true)}
-      onMouseOut={() => setHover(false)}
-      onFocus={() => undefined}
-      onBlur={() => undefined}
-    >
-      {scenarioHour}
-      {scenarioStatus}
-      {scenarioDetails}
-      {showMenu && <ScenarioMenu onContextMenuClick={onContextMenuClick} menu={menu} scenario={scenario} />}
+    <div style={{ display: 'flex', alignItems: 'center' }}>
+      <div
+        className={classes.scenario}
+        onMouseOver={() => setHover(true)}
+        onMouseOut={() => setHover(false)}
+        onFocus={() => undefined}
+        onBlur={() => undefined}
+        onClick={() => onClick(scenario)}
+      >
+        {scenarioHour}
+        {scenarioStatus}
+        {scenarioDetails}
+      </div>
+      {showMenu && (
+        <ScenarioMenu
+          onContextMenuClick={onContextMenuClick}
+          onClick={() => onClick(scenario)}
+          menu={menu}
+          scenario={scenario}
+        />
+      )}
     </div>
   );
 };