+++ packages/react-components/src/api/Scenarios/ScenariosApi.ts
@@ -1,5 +1,6 @@
 import { tap } from 'rxjs/operators';
 import { DataSource } from '../../api/types';
+import { ScenarioItem } from '../../Scenarios/ScenarioItem/ScenarioItem';
 import { fetchUrl } from '../helpers';
 
 /**
@@ -36,7 +37,7 @@ const fetchScenarios = (dataSource: DataSource, token: string) => {
     additionalHeaders: {
       Authorization: `Bearer ${token}`,
     },
-  }).pipe(tap((res) => console.log('List of scenarios fetched', res)));
+  });
 };
 
 /**
@@ -71,13 +72,27 @@ const fetchScenariosByDate = (dataSource: DataSource, token: string) => {
  * @param token
  * @param id
  */
-const deleteScenario = (dataSource: DataSource, token: string, id: any) =>
-  fetchUrl(`${dataSource.host}/api/scenarios/${dataSource.connection}/${id}`, {
-    method: 'DELETE',
-    additionalHeaders: {
-      Authorization: `Bearer ${token}`,
-    },
-  }).pipe(tap((res) => console.log('scenario deleted', res)));
+const deleteScenario = (dataSource: DataSource, token: string, scenario: any, softDelete = false) => {
+  if (softDelete) {
+    scenario.data.deleted = true;
+    scenario.data = JSON.stringify(scenario.data);
+
+    return fetchUrl(`${dataSource.host}/api/scenarios/${dataSource.connection}`, {
+      method: 'PUT',
+      additionalHeaders: {
+        Authorization: `Bearer ${token}`,
+      },
+      body: JSON.stringify(scenario),
+    });
+  } else {
+    return fetchUrl(`${dataSource.host}/api/scenarios/${dataSource.connection}/${scenario}`, {
+      method: 'DELETE',
+      additionalHeaders: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+  }
+};
 
 /**
  * /api/scenarios/{connectionId}
@@ -93,7 +108,7 @@ const postScenario = (dataSource: DataSource, token: string, scenario: any) =>
       Authorization: `Bearer ${token}`,
     },
     body: JSON.stringify(scenario),
-  }).pipe(tap((res) => console.log('scenario posted', res)));
+  });
 
 /**
  * /api/scenarios/{connectionId}
@@ -109,6 +124,6 @@ const updateScenario = (dataSource: DataSource, token: string, scenario: any) =>
       Authorization: `Bearer ${token}`,
     },
     body: JSON.stringify(scenario),
-  }).pipe(tap((res) => console.log('scenario updated', res)));
+  });
 
 export { fetchScenario, fetchScenarios, fetchScenariosByDate, deleteScenario, postScenario, updateScenario };